rename parsedConfig -> context
